+++ src/app/shared/components/modal/modal.component.ts
@@ -1,4 +1,5 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, Input, OnInit, Output } from '@angular/core';
+import { EventEmitter } from 'stream';
 
 @Component({
   selector: 'app-modal',
@@ -7,9 +8,25 @@ import { Component, OnInit } from '@angular/core';
 })
 export class ModalComponent implements OnInit {
 
+  @Input() title!: string;
+  @Input() description!: string;
+  @Input() image!: string;
+  @Input() isComic!: boolean;
+  @Input() id!: number;
+  @Output() data = new EventEmitter();
+  modal: boolean = false;
+
   constructor() { }
 
   ngOnInit(): void {
+  }
+
+  open() {
+    this.modal = true;
+  }
+
+  close() {
+    this.modal = false;
   }
 
 }